Correctly infers records even with variable shadowing
