+++ test/Test/Repl.hs
@@ -9,6 +9,7 @@ where
 
 import Data.Either (isLeft, isRight)
 import Data.Functor (($>))
+import qualified Data.Map as M
 import Data.Text (Text)
 import Language.Mimsa.Interpreter
 import Language.Mimsa.Printer
@@ -80,7 +81,7 @@ spec =
         result
           `shouldBe` Right
             ( MTFunction mempty (unknown 5) (unknown 5),
-              MyLambda mempty (named "i") (MyVar mempty (named "i"))
+              MyLambda mempty (named "i") (MyVar mempty (tvFree 1))
             )
       it "let prelude = ({ id: (\\i -> i) }) in prelude.id(1)" $ do
         result <- eval stdLib "let prelude = ({ id: (\\i -> i) }) in prelude.id(1)"
@@ -471,8 +472,28 @@ spec =
         result <- eval stdLib "\\a -> let one = a.one; let two = a.two; a"
         result
           `shouldSatisfy` isRight
-      xit "\\a -> let one = a.one in \\a -> let two = a.two in a" $ do
-        result <- eval stdLib "\\a -> let one = a.one in \\a -> let two = a.two in a.one"
+      it "\\a -> let one = a.one; \\a -> let two = a.two in a.one" $ do
+        result <- eval stdLib "\\a -> let one = a.one; \\a -> let two = a.two in a.one"
         -- here the two a's should be different types due to shadowing
-        result
-          `shouldSatisfy` isLeft
+        -- but even knowing the second a is different it can infer stuff
+        fst <$> result
+          `shouldBe` Right
+            ( MTFunction
+                mempty
+                ( MTRecord
+                    mempty
+                    (M.singleton (mkName "one") (MTVar mempty (tvFree 1)))
+                )
+                ( MTFunction
+                    mempty
+                    ( MTRecord
+                        mempty
+                        ( M.fromList
+                            [ (mkName "one", MTVar mempty (tvFree 2)),
+                              (mkName "two", MTVar mempty (tvFree 3))
+                            ]
+                        )
+                    )
+                    (MTVar mempty (tvFree 2))
+                )
+            )
