+++ test/golden/PrettyPrint/6a1bbccec9e7c694ed978aed3b4d00639ea425edf6c56b5450fadfc54f8ecea2.mimsa
@@ -1,0 +1,5 @@
+\a ->
+   let one = a.one;
+   \a ->
+      let two = a.two;
+      a.other
