+++ test/golden/StoreExpr/6a1bbccec9e7c694ed978aed3b4d00639ea425edf6c56b5450fadfc54f8ecea2.json
@@ -1,0 +1,1 @@
+{"storeExpression":{"tag":"MyLambda","contents":[[],"a",{"tag":"MyLet","contents":[[],"one",{"tag":"MyRecordAccess","contents":[[],{"tag":"MyVar","contents":[[],"a"]},"one"]},{"tag":"MyLambda","contents":[[],"a",{"tag":"MyLet","contents":[[],"two",{"tag":"MyRecordAccess","contents":[[],{"tag":"MyVar","contents":[[],"a"]},"two"]},{"tag":"MyRecordAccess","contents":[[],{"tag":"MyVar","contents":[[],"a"]},"other"]}]}]}]}]},"storeBindings":{},"storeTypeBindings":{}}
